pyproject.toml
[project]
name = "selfcal-omr"
version = "0.1.0"
description = "Self-calibrating OpenCV-based OMR library for bubble sheet detection."
authors = [
  { name="Hamed yazdan Nejat" }
]
dependencies = [
  "opencv-python",
  "numpy",
  "Pillow"
]
requires-python = ">=3.8"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["selfcal_omr*"]
exclude = ["outputs*", "tests*"]
